+++ src/coral_fans/functions/func/AutoItem.cpp
@@ -1,3 +1,4 @@
+#include "coral_fans/base/Mod.h"
 #include "ll/api/memory/Hook.h"
 #include "mc/nbt/CompoundTag.h"
 #include "mc/world/Container.h"
@@ -22,33 +23,52 @@ LL_TYPE_INSTANCE_HOOK(
     ::ItemUseMethod  itemUseMethod,
     bool             consumeItem
 ) {
+    if ((coral_fans::mod().getConfigDb()->get(std::format("functions.players.{}.autoitem", this->getUuid().asString()))
+         == "false"))
+        return origin(item, itemUseMethod, consumeItem);
     std::string name = item.getTypeName();
-    origin(item, itemUseMethod, consumeItem);
-    if (item == ItemStack::EMPTY_ITEM()) {
+    if (name.ends_with("_shulker_box")) {
+        if (!this->mInventory->mInventory) return;
         Container& inv          = *this->mInventory->mInventory;
         int        size         = inv.getContainerSize();
         int        selectedSlot = this->getSelectedItemSlot();
         for (int i = 0; i < size; i++) {
             if (i == selectedSlot) continue;
             const ItemStack& itemi = inv.getItem(i);
-            if (itemi.getTypeName() == name) {
-                item = itemi;
+            if (item.matchesItem(itemi)) {
                 inv.setItem(i, ItemStack::EMPTY_ITEM());
                 return;
             }
-            if (itemi.getTypeName().ends_with("_shulker_box")) {
-                auto tag = itemi.save(*SaveContextFactory::createCloneSaveContext());
-                if (!tag->contains("tag")) continue;
-                auto list  = (*tag)["tag"]["Items"].get<ListTag>();
-                int  _size = list.size();
-                for (int _i = 0; _i < _size; _i++) {
-                    auto itemTag = list.getCompound(_i);
-                    if ((*itemTag)["Name"].get<StringTag>() == name) {
-                        item = ItemStack::fromTag(*itemTag);
-                        list.erase(list.begin() + _i);
-                        (*tag)["tag"]["Items"] = list;
-                        inv.setItem(i, ItemStack::fromTag(*tag));
-                        return;
+        }
+        origin(item, itemUseMethod, consumeItem);
+    } else {
+        origin(item, itemUseMethod, consumeItem);
+        if (item == ItemStack::EMPTY_ITEM()) {
+            Container& inv          = *this->mInventory->mInventory;
+            int        size         = inv.getContainerSize();
+            int        selectedSlot = this->getSelectedItemSlot();
+            for (int i = 0; i < size; i++) {
+                if (i == selectedSlot) continue;
+                const ItemStack& itemi = inv.getItem(i);
+                if (itemi.getTypeName() == name) {
+                    item = itemi;
+                    inv.setItem(i, ItemStack::EMPTY_ITEM());
+                    return;
+                }
+                if (itemi.getTypeName().ends_with("_shulker_box")) {
+                    auto tag = itemi.save(*SaveContextFactory::createCloneSaveContext());
+                    if (!tag || !tag->contains("tag") || !(*tag)["tag"].contains("Items")) continue;
+                    auto list  = (*tag)["tag"]["Items"].get<ListTag>();
+                    int  _size = list.size();
+                    for (int _i = 0; _i < _size; _i++) {
+                        auto itemTag = list.getCompound(_i);
+                        if ((*itemTag)["Name"].get<StringTag>() == name) {
+                            item = ItemStack::fromTag(*itemTag);
+                            list.erase(list.begin() + _i);
+                            (*tag)["tag"]["Items"] = list;
+                            inv.setItem(i, ItemStack::fromTag(*tag));
+                            return;
+                        }
                     }
                 }
             }