+++ src/coral_fans/functions/other/Shortcuts.h
@@ -1,8 +1,0 @@
-#pragma once
-
-namespace coral_fans::functions {
-
-void registerShortcutsListener();
-void registerShortcutsCommand();
-
-} // namespace coral_fans::functions