+++ src/coral_fans/functions/slime/Slime.h
@@ -11,12 +11,13 @@ namespace coral_fans::functions {
 class SlimeManager {
 private:
     std::unordered_map<ChunkPos, bsci::GeometryGroup::GeoId> mParticleMap;
-    bool                                                     mShow;
 
 public:
-    inline void setShow(bool enable) { this->mShow = enable; }
-    void        tick();
-    void        remove();
+    bool mShow;
+
+public:
+    void tick();
+    void remove();
 };
 
 } // namespace coral_fans::functions